requirements.txt
certifi==2019.9.11
cffi==1.13.1
cryptography==2.8
future==0.18.2
numpy==1.17.3
pandas==0.25.2
pickleDB==0.9.2
pycparser==2.19
python-dateutil==2.8.0
python-telegram-bot==12.2.0
pytz==2019.3
scipy==1.3.1
six==1.12.0
summa==1.2.0
tornado==6.0.3
python-socketio==4.5.0
aiohttp==3.6.2